Move ENA utils to utils subdir
